==> 3-dev_datascience/1-experiment_train.ipynb <==
## 🐈 Building the model

In the journey of building a machine learning model, one of the first decisions is choosing the right type of model—predictive or generative. Predictive models focus on forecasting outcomes based on input data, while generative models aim to learn the underlying distribution of data to generate new samples.

Our usecase is categorized under predictive machine learning. There are many different ways to build a predictive model. For our use case, we chose neural networks. It has the ability to generalize better, handle complex patterns, and more expressive.


## 🐠 Install & Import packages

Again, we will need to install and import packages as we develop our notebook.

!pip -q install keras "tensorflow==2.15.1" "tf2onnx" "onnx" "seaborn" "onnxruntime"

import pandas as pd
from sklearn.preprocessing import StandardScaler, MinMaxScaler, OneHotEncoder, LabelEncoder
from sklearn.model_selection import train_test_split
from pathlib import Path
import pickle
import os

import numpy as np

from keras.models import Sequential
from keras.layers import Dense, Dropout, BatchNormalization, Activation
import tf2onnx
import onnx
import tensorflow as tf

from sklearn.metrics import confusion_matrix
import numpy as np
import pickle
import seaborn as sns
from matplotlib import pyplot as plt
import onnxruntime as rt

# 📦 Load Data

We again load our two datasets, merge them, drop the NA columns just like before and select the input and output data.

Input data (X) is the feature matrix that contains the characteristics of each song.

Output data (y)is the target variable the model is trying to predict. In this case, y is the 'country' column which represents the country where the song is popular. The model will learn to predict the country based on the song features in X.


song_properties = pd.read_parquet('https://github.com/rhoai-mlops/jukebox/raw/refs/heads/main/1-data_prep/song_properties.parquet')
song_rankings = pd.read_parquet('https://github.com/rhoai-mlops/jukebox/raw/refs/heads/main/1-data_prep/song_rankings.parquet')
song_properties.keys()

song_rankings = song_rankings.dropna()

X = song_rankings.merge(song_properties, on='spotify_id', how='left')
X = X[['is_explicit', 'duration_ms', 'danceability', 'energy', 'key', 'loudness', 'mode', 'speechiness', 'acousticness', 'instrumentalness', 'liveness', 'valence', 'tempo']]
y = song_rankings['country']

label_encoder = LabelEncoder()
y_encoded = label_encoder.fit_transform(y)
y_one_hot = tf.keras.utils.to_categorical(y_encoded)

# Split the data into training and testing sets so you have something to test the trained model with.
X_train, X_test, y_train, y_test = train_test_split(X, y_one_hot, test_size = 0.2, shuffle = False)
X_train, X_val, y_train, y_val = train_test_split(X_train,y_train, test_size = 0.2, stratify = y_train)

# Scale the data to remove mean and have unit variance. The data will be between -1 and 1, which makes it a lot easier for the model to learn than random (and potentially large) values.
# It is important to only fit the scaler to the training data, otherwise you are leaking information about the global distribution of variables (which is influenced by the test set) into the training set.
scaler = MinMaxScaler()
scaled_x_train = scaler.fit_transform(X_train.values)

# 🛟 Prepare to save the model

Before we start building neural network and training the model, let's prepare the environment to store the resulting artifacts. 

We  need to follow a specific structure due to OpenVino. OpenVino is an open-source toolkit that is designed to facilitate the optimization and deployment of models. This is what we are going to use to serve our models from OpenShift AI. 

Therefore we need to store our model artifacts in an S3 buckets with folder called models/<model-name>/<version>/

# create local directories to save the model artifacts before starting building neural network and training the model
Path("models/jukebox/1/artifacts").mkdir(parents=True, exist_ok=True)

with open("models/jukebox/1/artifacts/scaler.pkl", "wb") as handle:
    pickle.dump(scaler, handle)

with open("models/jukebox/1/artifacts/label_encoder.pkl", "wb") as handle:
    pickle.dump(label_encoder, handle)

# 🚀 Build the model

The below piece of code is like creating a smart helper that learns to guess which country might like a song based on its features. We start by building different layers in our helper that help it understand all the different characteristics of the song. Each layer gets better at figuring out patterns, just like how we learn from examples. 

At the end, our little helper uses what it learned to predict the country that would enjoy the song the most. 

Finally, we check how well our helper is doing at making these guesses!

model = Sequential()
model.add(Dense(32, activation = 'relu', input_dim = len(X.columns)))
model.add(Dense(64))
model.add(Activation('relu'))
model.add(Dense(128))
model.add(Activation('relu'))
model.add(Dense(256))
model.add(Activation('relu'))
model.add(Dense(y_one_hot.shape[1], activation = 'sigmoid'))
model.compile(optimizer='adam',loss='categorical_crossentropy',metrics=['accuracy', 'Precision', 'Recall'])
model.summary()

# 🏃 Train the Model

Now we train our smart helper to predict which country might like a song based on its features. We set it to learn from the training data for 2 rounds (epochs). During each round, it looks at the song characteristics (scaled_x_train) and the country labels (y_train), while also checking its guesses against a separate set of song data (X_val) to see how well it’s doing (validation_data). Remember we split the data into two in an above cell. That was the reason :)

Once the training is finished, we print a message to let us know that our model is ready to make predictions!

epochs = 2
history = model.fit(scaled_x_train, y_train, epochs=epochs, \
                    validation_data=(scaler.transform(X_val.values),y_val), \
                    verbose = True)
print("Training of model is complete")

# 🫡 Save the Model

Here we convert our trained song prediction model into a format that can be served via OpenShift AI.

model_proto, _ = tf2onnx.convert.from_keras(model)
os.makedirs("models/jukebox/1", exist_ok=True)
onnx.save(model_proto, "models/jukebox/1/model.onnx")

# 🔥 Load the Model for Testing

Here we load the model to predict which country might like a song. We open a session to the model and feed in the test data (X_test). The model outputs predictions, and identifies the countries.

The accuracy is calculated by comparing the predicted indices to the actual ones as we have the actual answers in our data. That's how we get the accuracy metrics.

We also create a confusion matrix to visualize the prediction results, using a heatmap to show how well the model's predictions match the actual labels. The darker the line, the closer we get to the actual answer.

sess = rt.InferenceSession("models/jukebox/1/model.onnx", providers=rt.get_available_providers())
input_name = sess.get_inputs()[0].name
output_name = sess.get_outputs()[0].name
y_pred_temp = sess.run([output_name], {input_name: scaler.transform(X_test.values).astype(np.float32)}) 
y_pred_temp = y_pred_temp[0]
y_pred_argmax = np.argmax(y_pred_temp, axis=1)

y_test_argmax = np.argmax(y_test, axis=1)

accuracy = np.sum(y_pred_argmax == y_test_argmax) / len(y_pred_argmax)
print("Accuracy: " + str(accuracy))

c_matrix = confusion_matrix(y_test_argmax,y_pred_argmax)
ax = sns.heatmap(c_matrix, cmap='Blues')
ax.set_xlabel("Prediction")
ax.set_ylabel("Actual")
ax.set_title('Confusion Matrix')
plt.show()

And now we need to save the model in our S3 bucket to make it available outside of this Notebook. So please open up the [2-save_model.ipynb](2-save_model.ipynb) :)
